Modulize jsonToTable
Change file name from jsonToTable.py to json_to_table.py and make
it as a module.
We can use it by 'from json_to_table import parse_json'.
